separate vxlan and ivxlan code to make the changes upstreamable
1. All the ivxlan changes migrated to vport-ivxlan.c that can be
   compiled as a separate driver.
2. Kernel vxlan driver would provide a registration api for enabling
   vxlan extensions (like ivxlan)
3. ivxlan driver registers parse and build functions for ivxlan extensions.

sample parse callstack when processing ivxlan traffic
    ovs_dp_process_packet
    ovs_vport_receive
    ivxlan_udp_encap_parse_hdr
    vxlan_udp_encap_recv
    udp_queue_rcv_skb
    __udp4_lib_rcv
    udp_rcv

sample parse callstack when processing plain traffic for ivxlan encap
     ivxlan_construct_hdr
     ivxlan_tnl_send
     ovs_vport_send
     do_output
     do_execute_actions
     ovs_execute_actions
     ovs_dp_process_packet

tested both encap and decap of ivxlan traffic and verified via
wireshark the sepg and source policy applied bits are set.

+++ lib/match.c
@@ -334,7 +334,7 @@ match_set_tun_ivxlan_spa_masked(struct match *match,
 void
 match_set_tun_ivxlan_spa(struct match *match, uint8_t ivxlan_spa)
 {
-    match_set_tun_ivxlan_spa_masked(match, ivxlan_spa, UINT16_MAX);
+    match_set_tun_ivxlan_spa_masked(match, ivxlan_spa, UINT8_MAX);
 }
 
 void
